Add name, description, and credits fields to editor screen
Moved the name, description, and credit editors into the project editor screen. Users can now view and update all the fields at the same time.
