+++ assets/js/api/CustomTranslationApi.js
@@ -2,7 +2,11 @@ import $ from 'jquery'
 
 export class CustomTranslationApi {
   constructor (programSection) {
-    this.programSection = programSection
+    if (programSection === 'credits') {
+      this.programSection = 'credit'
+    } else {
+      this.programSection = programSection
+    }
   }
 
   getCustomTranslation (programId, language, successCallback, errorCallback = () => {}) {
@@ -15,20 +19,38 @@ export class CustomTranslationApi {
   }
 
   deleteCustomTranslation (programId, language, successCallback, errorCallback) {
-    $.ajax({
-      url: '../translate/custom/project/' + programId + '?field=' + this.programSection + '&language=' + language,
-      type: 'delete',
-      success: successCallback,
-      error: errorCallback
+    const self = this
+    return new Promise(function (resolve, reject) {
+      $.ajax({
+        url: '../translate/custom/project/' + programId + '?field=' + self.programSection + '&language=' + language,
+        type: 'delete',
+        success: function (data) {
+          successCallback(language)
+          resolve()
+        },
+        error: function (error) {
+          errorCallback(error)
+          reject(error.status)
+        }
+      })
     })
   }
 
   saveCustomTranslation (programId, text, language, successCallback, errorCallback) {
-    $.ajax({
-      url: '../translate/custom/project/' + programId + '?field=' + this.programSection + '&text=' + text + '&language=' + language,
-      type: 'put',
-      success: successCallback,
-      error: errorCallback
+    const self = this
+    return new Promise(function (resolve, reject) {
+      $.ajax({
+        url: '../translate/custom/project/' + programId + '?field=' + self.programSection + '&text=' + text + '&language=' + language,
+        type: 'put',
+        success: function (data) {
+          successCallback(language)
+          resolve()
+        },
+        error: function (error) {
+          errorCallback(error)
+          reject(error.status)
+        }
+      })
     })
   }
 }
