+++ assets/js/custom/ProgramName.js
@@ -1,34 +1,10 @@
 import $ from 'jquery'
 
-export function ProgramName (programId, usersLanguage, myProgram, editor, editorConfig, customTranslationApi) {
+export function ProgramName (programId, usersLanguage, myProgram, customTranslationApi, editor) {
   const name = $('#name')
-  const editNameButton = $('#edit-name-button')
-  const descriptionCreditsContainer = $('#description-credits-container')
-  const showMoreToggle = $('#descriptionShowMoreToggle')
+  const editProgramButton = $('#edit-program-button')
 
-  if (myProgram) {
-    const descriptionHeadline = $('#description-headline')
-
-    editNameButton.on('click', () => {
-      editNameButton.hide()
-      descriptionCreditsContainer.hide()
-      descriptionHeadline.hide()
-      showMoreToggle.addClass('d-none')
-
-      editor.show(
-        editorConfig,
-        name.text().trim(),
-        closeNameEditor
-      )
-    })
-
-    function closeNameEditor () {
-      editNameButton.show()
-      descriptionCreditsContainer.show()
-      descriptionHeadline.show()
-      handleShowMore()
-    }
-  } else {
+  if (!myProgram) {
     customTranslationApi.getCustomTranslation(
       programId,
       usersLanguage.substring(0, 2),
@@ -40,9 +16,7 @@ export function ProgramName (programId, usersLanguage, myProgram, editor, editor
     }
   }
 
-  function handleShowMore () {
-    if (descriptionCreditsContainer.height() === 200 || descriptionCreditsContainer.height() > 300) {
-      showMoreToggle.removeClass('d-none')
-    }
-  }
+  editProgramButton.on('click', () => {
+    editor.show()
+  })
 }
