+++ assets/styles/custom/program.scss
@@ -620,9 +620,7 @@ input[type="number"] {
   z-index: 100;
   width: 100vw;
   height: calc(100vh - 50px);
-  padding-top: 1rem;
-  padding-right: math.div($grid-gutter-width, 2);
-  padding-left: math.div($grid-gutter-width, 2);
+  padding: 1rem math.div($grid-gutter-width, 2) 1.5rem;
   overflow-y: auto;
   background: #fff;
 }
