+++ templates/Program/program.html.twig
@@ -29,8 +29,8 @@
 
         {% if app.user and my_program %}
           <span>
-            <i id="edit-name-button" class="ms-3 material-icons catro-icon-button align-bottom"
-              data-bs-toggle="tooltip" title="{{ 'programs.editName'|trans({}, 'catroweb') }}">
+            <i id="edit-program-button" class="ms-3 material-icons catro-icon-button align-bottom"
+              data-bs-toggle="tooltip" title="{{ 'programs.editProject'|trans({}, 'catroweb') }}">
               edit
             </i>
           </span>
@@ -115,6 +115,8 @@
   {% include 'Program/program_reactions_modal.html.twig' %}
 
   {% include 'Program/program_description.html.twig' %}
+
+  {% include 'Program/program_editor.html.twig' %}
 
   {% include 'Program/program_tags.html.twig' %}
 
@@ -253,28 +255,19 @@
   ></div>
   <div class="js-project-description-credits"
        data-project-id="{{ program.id }}"
-       data-max-name-length={{ max_name_size }}
-       data-max-length={{ max_description_size }}
        data-has-description={% if program.description %} true {% else %} false {% endif %}
        data-has-credits={% if program.credits %} true {% else %} false {% endif %}
        data-trans-more-info="{{ "more-information"|trans({}, "catroweb") }}"
        data-trans-less-info="{{ "less-information"|trans({}, "catroweb") }}"
-       data-trans-name="{{ "name"|trans({}, "catroweb") }}"
-       data-trans-description="{{ "description"|trans({}, "catroweb") }}"
-       data-trans-notes-and-credits="{{ "credits"|trans({}, "catroweb") }}"
-       data-trans-translation-saved="{{ "programs.translationSaved"|trans({}, "catroweb") }}"
-       data-trans-translation-deleted="{{ "programs.translationDeleted"|trans({}, "catroweb") }}"
-       data-trans-close-name-editor="{{ "programs.closeNameEditor"|trans({}, "catroweb") }}"
-       data-trans-close-description-editor="{{ "programs.closeDescriptionEditor"|trans({}, "catroweb") }}"
-       data-trans-close-credits-editor="{{ "programs.closeCreditsEditor"|trans({}, "catroweb") }}"
+       data-trans-translation-updated="{{ "programs.translationUpdated"|trans({}, "catroweb") }}"
+       data-trans-close-editor="{{ "programs.closeEditor"|trans({}, "catroweb") }}"
        data-trans-close-editor-prompt="{{ "programs.closeEditorPrompt"|trans({}, "catroweb") }}"
        data-trans-save="{{ "programs.save"|trans({}, "catroweb") }}"
        data-trans-discard="{{ "programs.btnDiscard"|trans({}, "catroweb") }}"
        data-trans-cancel="{{ "cancel"|trans({}, "catroweb") }}"
        data-trans-keep="{{ "programs.btnKeep"|trans({}, "catroweb") }}"
        data-trans-default="{{ "programs.default"|trans({}, "catroweb") }}"
-       data-trans-save-on-language-change="{{ "programs.saveOnLanguageChange"|trans({}, "catroweb") }}"
-       data-trans-notes-and-credits-description="{{ "notesAndCreditsDescription"|trans({}, "catroweb") }}"
+       data-trans-save-on-language-change="{{ "programs.saveOnLanguageChange"|trans({}, "catroweb") }}"\
        data-path-edit-program-name="{{ path('edit_program_name', {'id': program.id}) }}"
        data-path-edit-program-description="{{ path('edit_program_description', {'id': program.id}) }}"
        data-path-edit-program-credits="{{ path('edit_program_credits', {'id': program.id}) }}"
